backend/requirements.txt
aiosmtplib==2.0.2
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.4
fastapi-cli==0.0.5
fastapi-mail==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.11
passlib==1.7.4
playwright==1.41.1
psycopg==3.2.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-extra-types==2.10.0
pydantic-settings==2.6.1
pydantic_core==2.23.4
pyee==11.0.1
Pygments==2.18.0
PyJWT==2.9.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
PyYAML==6.0.2
rich==13.9.4
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typer==0.13.0
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==14.1
